Revert "[WIP] Nt/post display optimise"
