+++ src/components/quickReplyModal/quickReplyModalContent.tsx
@@ -36,7 +36,8 @@ import { Draft } from '../../redux/reducers/cacheReducer';
 import { RootState } from '../../redux/store/store';
 
 import { PointActivityIds } from '../../providers/ecency/ecency.types';
-import { useUserActivityMutation, postQueries } from '../../providers/queries';
+import { useUserActivityMutation } from '../../providers/queries/pointQueries';
+import { usePostsCachePrimer } from '../../providers/queries';
 
 export interface QuickReplyModalContentProps {
   selectedPost?: any;
@@ -49,7 +50,7 @@ export const QuickReplyModalContent = forwardRef(
     const intl = useIntl();
     const dispatch = useDispatch();
     const userActivityMutation = useUserActivityMutation();
-    const postsCachePrimer = postQueries.usePostsCachePrimer();
+    const postsCachePrimer = usePostsCachePrimer();
 
     const inputRef = useRef<RNTextInput | null>(null);
