--- src/providers/queries/postQueries/index.ts
@@ -1,3 +1,0 @@
-import * as postQueries from './postQueries';
-
-export { postQueries };
